Reorganize character-reading switch
Two switches are now used: one outer switch, which has special cases for
the complex looping characters, and one inner switch, for the other simple
and contextless operations. This removes the repetition in the parsing
of the simple operations, and makes the complex loop parsing clearer.

Minor changes to the comments in the loop-parsing sections, as well.