Add composite brainfuck::error type
On the library side of things, it is easier to remember that brainfuck
functions can return an error type, rather than have to look up the
documentation for load() and run() separately. Even if the names are
obvious once you know them, it is a burden. This design is in line with
other modules, such as io.